Switch to simpler API.
